Create ExcelReader class with basic parsing method

+++ src/main/java/org/example/model/WorkerTimeSheet.java
@@ -11,5 +11,5 @@ import java.time.LocalDate;
 public class WorkerTimeSheet {
     private LocalDate date;
     private String task;
-    private float time;
+    private double time;
 }
